# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.DocumentListResult:
    properties:
      documents:
        items:
          $ref: '#/definitions/models.Document'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.LoginInput:
    properties:
      password:
        example: test123
        minLength: 6
        type: string
      username:
        example: test
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  dto.UpdateDocumentInput:
    properties:
      description:
        type: string
      file_name:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
  models.Document:
    properties:
      created_at:
        type: string
      description:
        type: string
      doc_num:
        type: integer
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      knowledge_base:
        $ref: '#/definitions/models.KnowledgeBase'
      knowledge_base_id:
        type: string
      name:
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.JSONMap:
    additionalProperties: true
    type: object
  models.KnowledgeBase:
    properties:
      created_at:
        type: string
      dataset_id:
        type: string
      description:
        type: string
      doc_num:
        type: integer
      document_count:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      tags:
        $ref: '#/definitions/models.JSONMap'
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      documents:
        description: 关联
        items:
          $ref: '#/definitions/models.Document'
        type: array
      email:
        type: string
      id:
        type: string
      nickname:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 用于用户增删改查的 RESTful API
  title: 用户管理系统 API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户凭据并返回JWT令牌
      parameters:
      - description: 登录凭据
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功返回令牌
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 认证失败
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 用户登录
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 刷新JWT令牌以延长会话
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功返回新令牌
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: 令牌无效或过期
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 刷新令牌
      tags:
      - 认证
  /documents:
    get:
      consumes:
      - application/json
      description: 分页获取指定知识库下的所有文档
      parameters:
      - description: 知识库ID
        in: query
        name: knowledge_base_id
        required: true
        type: string
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文档列表
          schema:
            $ref: '#/definitions/dto.DocumentListResult'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 获取知识库下所有文档
      tags:
      - 文档模块
  /documents/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的文档
      parameters:
      - description: 文档ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 文档不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 删除文档
      tags:
      - 文档模块
    get:
      consumes:
      - application/json
      description: 根据ID获取文档的详细信息
      parameters:
      - description: 文档ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文档详情
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 文档不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 获取文档详情
      tags:
      - 文档模块
    put:
      consumes:
      - application/json
      description: 更新文档信息
      parameters:
      - description: 文档ID (UUID格式)
        in: path
        name: id
        required: true
        type: string
      - description: 更新输入
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDocumentInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的文档
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 文档不存在
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 更新文档
      tags:
      - 文档模块
  /documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文档到指定知识库
      parameters:
      - description: 知识库ID
        in: formData
        name: knowledge_base_id
        required: true
        type: string
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功的文档信息
          schema:
            $ref: '#/definitions/models.Document'
        "400":
          description: 请求参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: 上传文档
      tags:
      - 文档模块
...
securityDefinitions:
  Bearer:
    description: '请输入 ''Bearer '' + JWT 令牌，例如: "Bearer abcde12345"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"